+++ src/python/stup/core/core.py
@@ -15,9 +15,9 @@ from stup import config as Config
 from stup import packet as StupPacket
 from stup import utils as Utils
 from stup.state import machine as StateMachine
-from stup import twisted
+from stup import twistedutils
 from stup.ring import RingBuffer, RingWindow, ContainerItem
-from stup.twisted.deferred_deque import DeferredDeque
+from stup.twistedutils.deferred_deque import DeferredDeque
 
 class StupCore(object):
     class PacketItem(object):
@@ -60,8 +60,8 @@ class StupCore(object):
         self.last_recv_ack_id = -1
         self.last_recv_ack_cnt = 0
 
-        self.time_wheel_fine   = twisted.TimeWheel(255, Config.TIMER_FINE_GRANULARITY)
-        self.time_wheel_coarse = twisted.TimeWheel(10, Config.TIMER_COARSE_GRANULARITY)
+        self.time_wheel_fine   = twistedutils.TimeWheel(255, Config.TIMER_FINE_GRANULARITY)
+        self.time_wheel_coarse = twistedutils.TimeWheel(10, Config.TIMER_COARSE_GRANULARITY)
         self.tick_fine()
         self.tick_coarse()
