+++ src/python/stup/twistedutils/deferred_deque.py
