+++ src/python/stup/twistedutils/utils.py
